fix: mentor, cadet, bocalのレポジトリを別のフォルダーに分離する
