+++ src/v1/bocals/bocals.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Bocals } from 'src/domain/typeorm/entity/bocal.entity';
 import { BocalsService } from './bocals.service';
-import { BocalsRepository } from './bocals.repository';
+import { BocalsRepository } from './repository/bocals.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Bocals])],
