+++ src/v1/cadets/cadets.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { CadetsService } from './cadets.service';
 import { Cadets } from 'src/domain/typeorm/entity/cadets.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CadetsRepository } from './cadets.repository';
+import { CadetsRepository } from './repository/cadets.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Cadets])],
