+++ src/v1/mentors/repository/mentors.repository.ts
@@ -30,4 +30,37 @@ export class MentorsRepository {
 
     return mentor;
   }
+
+  async findByIntra(intraId: string): Promise<Mentors> {
+    let foundUser: Mentors;
+
+    try {
+      foundUser = await this.mentorsRepository.findOneBy({
+        intraId,
+      });
+    } catch (error) {
+      throw new ConflictException(error, process.env.CONFLICTEXCEPTION);
+    }
+
+    if (!foundUser) {
+      throw new NotFoundException(process.env.NOTFOUNDEXECEPTION);
+    }
+
+    return foundUser;
+  }
+
+  async createUser(intraId: string): Promise<Mentors> {
+    try {
+      const createdMentors: Mentors = this.mentorsRepository.create({
+        intraId: intraId,
+        isActive: false,
+      });
+
+      const updatedMentor = await this.mentorsRepository.save(createdMentors);
+
+      return updatedMentor;
+    } catch (error) {
+      throw new ConflictException(error, process.env.CONFLICTEXCEPTION);
+    }
+  }
 }
